pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioseeker"
version = "1.0.0"
description = "A tool for analyzing conservation of inhibitory codon pairs."
authors = [{name = "Facundo Martinez", email = "fxmartinez@proton.me"}]
dependencies = [
    "numpy",
    "pandas",
    "biopython",
    "matplotlib",
    "scipy",
]
requires-python = ">=3.9"

[project.scripts]
bioseeker = "bioseeker.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
